+++ src/br/edu/poo/gui/Locadora.java
@@ -1,0 +1,5 @@
+package br.edu.poo.gui;
+
+public class Locadora {
+
+}
