+++ src/br/edu/poo/objects/Caminhao.java
@@ -1,4 +1,4 @@
-package br.edu.poo_atividade_pratica;
+package br.edu.poo.objects;
 
 public class Caminhao extends Veiculo {
