+++ src/br/edu/poo/objects/Carro.java
@@ -1,4 +1,4 @@
-package br.edu.poo_atividade_pratica;
+package br.edu.poo.objects;
 
 public class Carro extends Veiculo {
