+++ src/br/edu/poo/objects/ListaVeiculo.java
@@ -1,16 +1,30 @@
-package br.edu.poo_atividade_pratica;
+package br.edu.poo.objects;
 
 public class ListaVeiculo implements IVeiculos {
     private Veiculo[] listaVeiculos;
     private int numVeiculos;
 
+    /**
+     * @param v veiculo do tipo Veiculo. 
+     * Se "v" não é nulo e é uma instância do tipo Veiculo, adiciona um veículo na lista de veículos.
+     * Caso contrário, não adiciona e retorna "Não é um veículo válido".
+     * !Inicializar a variável numVeiculos com "0"!
+     */
     @Override
     public void add(Veiculo v) {
         if (v != null && v instanceof Veiculo) {
             listaVeiculos[numVeiculos++] = v;
         }
+        System.err.println("Não é um veículo válido.");
     }
 
+    /**
+     * @param placa placa do veículo do tipo String.
+     * @return Veículo com a placa informada.
+     * @return "null" se a placa não for válida ou não for encontrada.
+     * Se a placa não for nula, procura o veículo na lista de veículos.
+     * Caso contrário, retorna "null".
+     */
     @Override
     public Veiculo get(String placa) {
         if (placa != null) {
@@ -23,6 +37,13 @@ public class ListaVeiculo implements IVeiculos {
         return null;
     }
 
+    /**
+     * @param placa placa do veículo do tipo String.
+     * @return Informações sobre o veículo com a placa informada.
+     * @return "Veículo não encontrado" se a placa não for válida ou não for encontrada.
+     * Se a placa não for nula, retorna informações sobre o veículo.
+     * Caso contrário, retorna "Veículo não encontrado".
+     */
     @Override
     public String getInfo(String placa) {
         if (placa != null) {
@@ -35,6 +56,12 @@ public class ListaVeiculo implements IVeiculos {
         return "Veiculo não encontrado";
     }
 
+    /**
+     * @return Informações sobre todos os veículos.
+     * @return "Não há veículos cadastrados" se não existir nenhum veículo.
+     * Se existir algum veículo, retorna informações sobre todos os veículos.
+     * Caso contrário, retorna "Não há veículos cadastrados".
+     */
     @Override
     public String getInfo() {
         if (listaVeiculos.length > 0) {
@@ -45,6 +72,12 @@ public class ListaVeiculo implements IVeiculos {
         return "Não há veículos cadastrados";
     }
 
+    /**
+     * @return Informações sobre todos os veículos.
+     * @return "Não há veículos cadastrados" se não existir nenhum veículo.
+     * Se existir algum veículo, retorna informações sobre todos os veículos.
+     * Caso contrário, retorna "Não há veículos cadastrados".
+     */
     @Override
     public String getResumoInfo() {
         if (listaVeiculos.length > 0) {
@@ -56,6 +89,13 @@ public class ListaVeiculo implements IVeiculos {
         return "Não há veículos cadastrados";
     }
 
+    /**
+     * @param placa placa do veículo do tipo String.
+     * @return "true" se a placa for válida e o veículo for encontrado.
+     * @return "false" se a placa não for válida ou não for encontrada.
+     * Se a placa não for nula, remove o veículo da lista de veículos e retorna true.
+     * Caso contrário, retorna false.
+     */
     @Override
     public boolean remove(String placa) {
         if (placa != null) {
@@ -71,6 +111,13 @@ public class ListaVeiculo implements IVeiculos {
         return false;
     }
 
+    /**
+     * @param placa placa do veículo do tipo String.
+     * @return "true" se a placa for válida e o veículo for encontrado.
+     * @return "false" se a placa não for válida ou não for encontrada.
+     * Se a placa não for nula, retorna true.
+     * Caso contrário, retorna false.
+     */
     @Override
     public boolean existe(String placa) {
         if (placa != null) {
